requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic-settings==2.12.0
requests==2.32.5
fpdf2==2.8.5
